add join_event_queue, finish minimal fix for #12
Fixes concurrency bug where App shutdown sequence used poll_events() to
try to dispatch all events remaining in queue. This is not correct
because poll_events() will poll and dispatch from the main thread, while
the event loop thread will also continue to do so. Since poll and
dispatch is not atomic, we can get incorrect interleavings of events,
causing, e.g. stop events to run in incorrect order, or a frame event to
run between stop events. The solution is to use Queue's task_done() and
join(), and just wait for the event loop thread to poll remaining
events.